Extract SRAM hot expert code into transforms/sram_experts.py + sram_slots.py
